Singleplayer World
Abstract game world is the scene, that can switch levels from one to other, using the methods `ascend` and `descend`. These methods can be overriden to make an unique behaviour. It also has a depth counter, that shows how deeper the player passed, and the signal `depth_changed` triggers on each descend/ascend.

Single player world is the game world implementation, that allows a player to control only one playable character on the scene.